registered origin constraints with /info
min_ttl, max_ttl, default_ttl, and max_cdn_file_size are now
exposed in the /info endpoint under the 'cdn_origin' key.

In order to keep only one place for setting the default values
in the code, the defaults for these variables is set in a constant.

A functional test was added to check that these values exist and
return something castable to an int.

Updated the sample config to reflect the default value set by the code.